consolidate duplicate function definitions
