+++ packages/frontend/editor-ui/src/components/ConsumedTokensDetails.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { formatTokenUsageCount } from '@/components/CanvasChat/future/utils';
+import { formatTokenUsageCount } from '@/components/RunDataAi/utils';
 import { useI18n } from '@/composables/useI18n';
 import { type LlmTokenUsageData } from '@/Interface';
 import { N8nText } from '@n8n/design-system';
